Moved files to ui subdirectory
