+++ frontend/app/src/main/AndroidManifest.xml
@@ -32,7 +32,7 @@
             android:value="${MAPS_API_KEY}" />
 
         <activity
-            android:name=".BaseActivity"
+            android:name=".ui.BaseActivity"
             android:exported="false" />
 
         <meta-data
@@ -40,11 +40,11 @@
             android:value="@integer/google_play_services_version" />
 
         <activity
-            android:name=".HomeActivity"
+            android:name=".ui.HomeActivity"
             android:exported="false"
             android:label="@string/title_activity_home" />
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
