+++ frontend/app/src/main/java/com/example/findmy/ui/AddPOIFragment.java
@@ -1,4 +1,4 @@
-package com.example.findmy;
+package com.example.findmy.ui;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.findmy.R;
 
 /**
  * A simple {@link Fragment} subclass.
