+++ frontend/app/src/main/java/com/example/findmy/ui/profile/ProfileFragment.java
@@ -12,7 +12,7 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
 
-import com.example.findmy.HomeActivity;
+import com.example.findmy.ui.HomeActivity;
 import com.example.findmy.databinding.FragmentProfileBinding;
