+++ frontend/app/src/main/res/layout/fragment_add_p_o_i.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddPOIFragment">
+    tools:context=".ui.AddPOIFragment">
 
 
     <EditText
