+++ frontend/app/src/main/res/layout/fragment_maps.xml
@@ -10,7 +10,7 @@
         android:name="com.google.android.gms.maps.SupportMapFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MapsFragment" />
+        tools:context=".ui.MapsFragment" />
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/newPOIButton"
