+++ frontend/app/src/main/res/navigation/mobile_navigation.xml
@@ -13,7 +13,7 @@
 
     <fragment
         android:id="@+id/navigation_map"
-        android:name="com.example.findmy.MapsFragment"
+        android:name="com.example.findmy.ui.MapsFragment"
         android:label="@string/title_map"
         tools:layout="@layout/fragment_maps" >
         <action
@@ -42,7 +42,7 @@
 
     <fragment
         android:id="@+id/navigation_add_poi"
-        android:name="com.example.findmy.AddPOIFragment"
+        android:name="com.example.findmy.ui.AddPOIFragment"
         android:label="@string/title_add_poi"
         tools:layout="@layout/fragment_add_p_o_i" />
 </navigation>
